+++ src/discharge_docs/api/app.py
@@ -1,6 +1,7 @@
 import json
 import logging
 import os
+import traceback
 from datetime import datetime
 from pathlib import Path
 
@@ -19,7 +20,11 @@ from discharge_docs.database.models import (
     ApiRequest,
     Base,
 )
-from discharge_docs.processing.processing import get_patient_file
+from discharge_docs.processing.deduce_text import apply_deduce
+from discharge_docs.processing.processing import (
+    get_patient_file,
+    process_data_metavision_dp,
+)
 from discharge_docs.prompts.prompt import (
     load_prompts,
     load_template_prompt,
@@ -27,6 +32,7 @@ from discharge_docs.prompts.prompt import (
 from discharge_docs.prompts.prompt_builder import PromptBuilder
 
 logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
 load_dotenv()
 
 with open(Path(__file__).parents[3] / "pyproject.toml", "rb") as f:
@@ -66,6 +72,9 @@ client = AzureOpenAI(
     azure_endpoint=os.getenv("AZURE_OPENAI_ENDPOINT", ""),
 )
 
+with open(Path(__file__).parent / "deployment_config.toml", "rb") as f:
+    deployment_config_dict = tomli.load(f)
+
 
 def get_db():
     try:
@@ -83,9 +92,9 @@ def check_api_key(api_key_header: str = Security(api_key_header)) -> str:
         raise HTTPException(403, "Unauthorized, Api Key not valid")
 
 
-@app.post("/update-discharge-docs")
-async def update_discharge_docs(
-    daily_updates: list[dict],
+@app.post("/process-and-generate-discharge-docs")
+async def process_and_generate_discharge_docs(
+    data: list[dict],
     db: Session = Depends(get_db),
     api_key: str = Depends(check_api_key),
 ) -> dict:
@@ -95,63 +104,93 @@ async def update_discharge_docs(
     Parameters
     ----------
     daily_updates : list[dict]
-        daily updates on patients in json-format,
+        raw export of patient data in json-format,
         needs at least the following fields:
-          end_id, department, date, value and description
+          TODO ADD FIELDS
     """
-    start_time = datetime.now()
-    daily_updates_df = pd.DataFrame.from_records(daily_updates)
-    daily_updates_df["date"] = pd.to_datetime(daily_updates_df["date"])
+    try:
+        start_time = datetime.now()
 
-    api_request = ApiRequest(
-        timestamp=start_time,
-        endpoint="/update-discharge-docs",
-        api_version=API_VERSION,
-    )
+        data_df = pd.DataFrame.from_records(data)
 
-    prompt_builder = PromptBuilder(
-        temperature=TEMPERATURE, deployment_name=deployment_name, client=client
-    )
+        processed_data = process_data_metavision_dp(data_df, nifi=True)
 
-    user_prompt, system_prompt = load_prompts()
+        processed_data = apply_deduce(processed_data, "value")
 
-    for enc_id in daily_updates_df["enc_id"].unique():
-        patient_file_string, patient_df = get_patient_file(daily_updates_df, enc_id)
-        department = patient_df["department"].values[0]
-        template_prompt = load_template_prompt(department)
-
-        token_length = prompt_builder.get_token_length(
-            patient_file_string, system_prompt, user_prompt, template_prompt
+        api_request = ApiRequest(
+            timestamp=start_time,
+            endpoint="/update-discharge-docs",
+            api_version=API_VERSION,
         )
 
-        logger.info(
-            f"Generating discharge doc for encounter {enc_id} "
-            f"and department {department}..."
+        prompt_builder = PromptBuilder(
+            temperature=deployment_config_dict["TEMPERATURE"],
+            deployment_name=deployment_config_dict["deployment_name"],
+            client=client,
         )
-        discharge_letter = prompt_builder.generate_discharge_doc(
-            patient_file=patient_file_string,
-            system_prompt=system_prompt,
-            user_prompt=user_prompt,
-            template_prompt=template_prompt,
-        )
+        user_prompt, system_prompt = load_prompts()
 
-        api_encounter = ApiEncounter(
-            encounter_hix_id=str(enc_id),
-            department=department,
-            input_token_length=token_length,
-        )
-        api_discharge_letter = ApiGeneratedDoc(
-            discharge_letter=json.dumps(discharge_letter)
-        )
-        api_encounter.generated_doc_relation = api_discharge_letter
-        api_request.encounter_relation.append(api_encounter)
+        for enc_id in processed_data["enc_id"].unique():
+            patient_file_string, patient_df = get_patient_file(processed_data, enc_id)
+            department = patient_df["department"].values[0]
+            template_prompt = load_template_prompt(department)
 
-    end_time = datetime.now()
-    runtime = (end_time - start_time).total_seconds()
-    api_request.runtime = runtime
-    api_request.response_code = 200
-    db.add(api_request)
-    db.commit()
+            token_length = prompt_builder.get_token_length(
+                patient_file_string, system_prompt, user_prompt, template_prompt
+            )
+
+            logger.info(
+                f"Generating discharge doc for encounter {enc_id} "
+                f"and department {department}..."
+            )
+
+            discharge_letter = prompt_builder.generate_discharge_doc(
+                patient_file=patient_file_string,
+                system_prompt=system_prompt,
+                user_prompt=user_prompt,
+                template_prompt=template_prompt,
+            )
+
+            discharge_letter.insert(
+                0,
+                {
+                    "Categorie": "Datum gegenereerd",
+                    "Beloop tijdens opname": "Deze brief is gegenereerd door AI op: "
+                    + str(datetime.now().strftime("%Y-%m-%d %H:%M")),
+                },
+            )
+
+            api_encounter = db.execute(
+                select(ApiEncounter).where(ApiEncounter.encounter_hix_id == str(enc_id))
+            ).scalar_one_or_none()
+
+            if not api_encounter:
+                api_encounter = ApiEncounter(
+                    encounter_hix_id=str(enc_id),
+                    patient_number=str(patient_df["patient_number"].values[0]),
+                    department=department,
+                )
+            else:
+                pass
+
+            api_discharge_letter = ApiGeneratedDoc(
+                discharge_letter=json.dumps(discharge_letter),
+                input_token_length=token_length,
+            )
+            api_encounter.generated_doc_relation.append(api_discharge_letter)
+            api_request.encounter_relation.append(api_encounter)
+
+        end_time = datetime.now()
+        runtime = (end_time - start_time).total_seconds()
+        api_request.runtime = runtime
+        api_request.response_code = 200
+        db.merge(api_request)
+        db.commit()
+
+    except Exception as e:
+        tb = traceback.format_exc()
+        logger.error(f"Error in update_discharge_docs: {e}\n{tb}")
+        raise HTTPException(500, "Internal Server Error") from e
     return {"message": "Success"}
 
 
@@ -167,6 +206,7 @@ async def remove_old_discharge_docs(
     ----------
     max_date : datetime
         The maximum date for which discharge documentation should be kept.
+        Example format: YYYY-MM-DDT00:00:00
     """
     start_time = datetime.now()
     api_request = ApiRequest(