+++ tests/test_api.py
@@ -7,13 +7,20 @@ from sqlalchemy.orm import Session
 from test_evaluate_docs import MockAzureOpenAI
 
 import discharge_docs.api.app as app
-from discharge_docs.api.app import remove_old_discharge_docs, update_discharge_docs
+from discharge_docs.api.app import (
+    process_and_generate_discharge_docs,
+    remove_old_discharge_docs,
+)
 
 
 class FakeExecute:
     def all(self):
         print("requested all results...")
         return []
+
+    def scalar_one_or_none(self):
+        print("requested scalar one or none...")
+        return None
 
     @property
     def rowcount(self):
@@ -46,12 +53,12 @@ class FakeDB(Session):
 
 @pytest.mark.asyncio
 async def test_api_discharge_docs(monkeypatch):
-    """Test the update_discharge_docs endpoint in the API."""
+    """Test the process_and_generate_discharge_docs endpoint in the API."""
     with open(Path(__file__).parent / "data" / "example_data.json", "r") as f:
         test_data = json.load(f)
 
     monkeypatch.setattr(app, "client", MockAzureOpenAI())
-    output = await update_discharge_docs(test_data, FakeDB())
+    output = await process_and_generate_discharge_docs(test_data, FakeDB())
     assert output == {"message": "Success"}